% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{blendColor}
\alias{blendColor}
\title{Blend Colors}
\usage{
blendColor(col1 = "#1b9e777F", col2 = "#7570b37F")
}
\arguments{
\item{col1}{Anything that can be interpreted by \code{rgb}
as a color.}

\item{col2}{Anything that can be interpreted by \code{rgb}
as a color.}
}
\value{
A \code{character} string with hex color, including
adjustment for transparency.
}
\description{
Generates a blended color from two transparent colors
}
\examples{

blendColor(col1 = "#1B9E777F", col2 = "#7570B37F")

}
\keyword{internal}
\keyword{plotting}
