% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolateRaster.R
\name{interpolateRaster}
\alias{interpolateRaster}
\title{Interpolate patchily sampled rasters}
\usage{
interpolateRaster(inputRaster)
}
\arguments{
\item{inputRaster}{An object of class \code{raster}}
}
\value{
An object of class raster
}
\description{
Uses thin plate spline regression from
\code{fields} package to interpolate missing raster values.
Note that this is only done in two dimensions, x and y.
}
\examples{
# Create sample raster
r <- raster(ncol=10, nrow=10)
values(r) <- 1:100

# Introduce a "hole"
values(r)[67:70] <- NA
interpolatedRaster <- interpolateRaster(r)

}
\keyword{dataPrep}
