% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{pointCompMap}
\alias{pointCompMap}
\title{Comparative point mapping}
\usage{
pointCompMap(
  occs1,
  occs2,
  spName,
  land = NA,
  occs1Col = "#bd0026",
  occs2Col = "#fd8d3c",
  agreeCol = "black",
  occs1Name = "Set 1",
  occs2Name = "Set 2",
  landCol = "gray",
  waterCol = "steelblue",
  ptSize = 1,
  ...
)
}
\arguments{
\item{occs1}{A \code{data.frame} with at least two columns
named "longitude" and "latitude" or that
can be coerced into this format.}

\item{occs2}{A \code{data.frame} with at least two columns
named "longitude" and "latitude" or that
can be coerced into this format.}

\item{spName}{A character string with the species
name to be used in the plot title.}

\item{land}{An optional coastline polygon shapefile
of type \code{sf} to provide geographic context for the
occurrence points.}

\item{landCol}{Color for land on map}

\item{waterCol}{Color for water on map}

\item{ptSize}{\code{numeric} value for \code{cex};
size of occurrence points on map.}

\item{...}{Additional optional arguments to pass to
\code{ggplot} initial plot object.}

\item{pt1Col}{Color for occurrence points on map}

\item{pt2Col}{Color for occurrence points on map}
}
\value{
A \code{data.frame} with two columns named "longitude"
and "latitude" or with names that were used when coercing
input data into this format.
}
\description{
A convenient wrapper around ggplot
to generate formatted plots comparing two sets of
occurrence point plots.
}
\note{
The x and y column names of \code{occs1} and \code{occs2}
must match.
}
\examples{
occs <- read.csv(system.file("extdata/Aphanopus_intermedius.csv",
                             package='voluModel'))
spName <- "Aphanopus intermedius"
pointMap(occs = occs, spName = spName, land = NA)

}
\seealso{
\code{\link[ggplot2:ggplot]{ggplot}}
}
\keyword{plotting}
