% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{pointMap}
\alias{pointMap}
\title{Point mapping}
\usage{
pointMap(occs, spName, world)
}
\arguments{
\item{occs}{A \code{data.frame} with at least two columns
named "longitude" and "latitude" or that
can be coerced into this format.}

\item{spName}{A character string with the species
name to be used in the plot title.}

\item{world}{An optional coastline shapefile to
provide geographic context for the occurrence points.}

\item{...}{Additional optional arguments to pass to
\code{ggplot} initial plot object.}
}
\value{
A \code{data.frame} with two columns named "longitude"
and "latitude" or with names that were used when coercing
input data into this format.
}
\description{
A convenient wrapper around ggplot
to generate formatted occurrence point plots.
}
\details{
The meat of this function is a special-case wrapper
around \code{getDynamicAlphaHull} from the \code{rangeBuilder} package.
The function documented here is especially useful in cases where
one wants to automatically generate training regions that overlap
the international date line. Regions that exceed the line are cut
and pasted into the appropriate hemisphere instead of being
deleted.
}
\examples{
library(rnaturalearth)

occs <- read.csv(system.file("extdata/", package='voluModel'))
world <- ne_countries(scale = "medium", returnclass = "sf")


}
\seealso{
\code{\link[ggplot2:ggplot]{ggplot}}
}
\keyword{backgroundSampling}
