% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{rasterCompFunction}
\alias{rasterCompFunction}
\title{Comparative raster mapping}
\usage{
rasterCompFunction(
  rast1 = NULL,
  rast2 = NULL,
  col1 = "red",
  col2 = "blue",
  rast1Name = "Set 1",
  rast2Name = "Set 2",
  colNeither = "white",
  colBoth = "purple",
  land = NA,
  landCol = "black",
  title = "A Raster Comparison",
  ...
)
}
\arguments{
\item{land}{An optional coastline polygon shapefile
of type \code{sf} to provide geographic context for the
occurrence points.}

\item{landCol}{Color for land on map}

\item{...}{Additional optional arguments to pass to
\code{spplot} initial plot object.}

\item{occs1}{A \code{data.frame} with at least two columns
named "longitude" and "latitude" or that
can be coerced into this format.}

\item{occs2}{A \code{data.frame} with at least two columns
named "longitude" and "latitude" or that
can be coerced into this format.}

\item{spName}{A character string with the species
name to be used in the plot title.}

\item{occs1Col}{Color for occurrence points on map}

\item{occs2Col}{Color for occurrence points on map}

\item{occs1Name}{An optional name for the first set
of occurrences, which will be color-coded to
\code{occs1Col} in the resulting plot.}

\item{occs2Name}{An optional name for the first set
of occurrences, which will be color-coded to
\code{occs2Col} in the resulting plot.}

\item{agreeCol}{Color for occurrence points shared
between \code{occs1} and \code{occs2}.}
}
\description{
A convenient wrapper around \code{spplot}
to generate formatted plots comparing two rasters.
}
\note{
The extents of \code{rast1} and \code{rast2}
must match.
}
\examples{

}
\seealso{
\code{\link[raster:spplot]{spplot}}
}
\keyword{plotting}
