% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolateRaster.R
\name{smoothRaster}
\alias{smoothRaster}
\title{Smooth rasters}
\usage{
smoothRaster(inputRaster, fast = F, ...)
}
\arguments{
\item{inputRaster}{An object of class \code{raster}}

\item{fast}{A logical operator. Setting to \code{T} triggers use
of \code{fastTps} instead of \code{Tps}.}

\item{...}{For any additional arguments passed to \code{Tps} or \code{fastTps}}
}
\value{
An object of class \code{RasterLayer}
}
\description{
Uses thin plate spline regression from
\code{fields} package to smooth raster values.
}
\details{
Original raster
are replaced with interpolated values. User has the
option of choosing \code{fastTps} to speed calculation,
but be advised that this is only an approximation
of a true thin plate spline.
}
\examples{
library(raster)
library(fields)
# Create sample raster
r <- raster(ncol=100, nrow=100)
values(r) <- 1:10000

# Introduce a "bubble"
values(r)[67:70] <- 10

# Smooth bubble with smoothRaster
interpolatedRaster <- smoothRaster(r)
fastInterp <- smoothRaster(r, fast = TRUE, aRange = 3.0)

}
\seealso{
\code{\link[fields]{Tps}}, \code{\link[fields]{fastTps}}
}
\keyword{dataPrep}
