% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{transpColor}
\alias{transpColor}
\title{Transparent Color}
\usage{
transpColor(color, percent = 50)
}
\arguments{
\item{color}{Anthing that can be interpreted by \code{rgb}
as a color.}

\item{percent}{A whole number between 0 and 100 specifying
how transparent the color should be.}
}
\value{
A \code{list} of length 2 with indices of the x and y
columns, respectively, followed by a message with a plain
text report of which columns were interpreted as x and y.
}
\description{
Generates transparent colors
}
\examples{

transpColor(color = "red", percent = 50)

}
\keyword{internal}
\keyword{plotting}
