% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{verticalSample}
\alias{verticalSample}
\title{Vertical sample}
\usage{
verticalSample(x = NULL, sampleAxis = "lon", axisValue = NA)
}
\arguments{
\item{x}{A multilayer \code{SpatRaster} object}

\item{sampleAxis}{Specifies whether a latitudinal ("lat") or
longitudinal ("long") transect is desired.}

\item{axisValue}{Numeric value specifying transect postion.}
}
\value{
A \code{data.frame} with values sampled across vertical
transect.
}
\description{
Samples data along a vertical transect
}
\note{
Only unprojected \code{SpatRaster} files are supported.
}
\keyword{internal}
\keyword{plotting}
