% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyzSample.R
\name{xyzSample}
\alias{xyzSample}
\title{Sampling from rasterBrick using 3D coordinates}
\usage{
xyzSample(occs, envBrick)
}
\arguments{
\item{occs}{A dataframe with at least three columns
named "longitude", "latitude", and "depth", or that
can be coerced into this format.}

\item{envBrick}{A \code{rasterBrick} object with
one environmental variable. Each layer represents
a depth slice. See Details for more information.}
}
\value{
Vector of environmental values equal in length
to number of rows of input \code{occs} \code{data.frame}.
}
\description{
Gets values at x,y,z occurrences from a
given 3D environmental variable brick
}
\details{
The \code{envBrick} \code{rasterBrick} object should
have numeric names that correspond with the beginning
depth of a particular depth slice. For example, one
might have three layers, one from 0 to 10m, one from
10 to 30m, and one from 30 to 100m. You would name the
layers in this brick \verb{names(envBrick) <- c(0, 10, 30}.
R will rename the layers "X0", "X10", and "X30". This
is expected behavior and \code{xyzSample} was written to
expect this. \code{xyzSample} identifies the layer name
that is closest to the depth layer value at a
particular X, Y coordinate, and samples the
environmental value at that 3D coordinate.
}
\examples{
library(raster)

# Create test raster
r1 <- raster(ncol=10, nrow=10)
values(r1) <- 1:100
r2 <- raster(ncol=10, nrow=10)
values(r2) <- c(rep(20, times = 50), rep(60, times = 50))
r3 <- raster(ncol=10, nrow=10)
values(r3) <- 8
envBrick <- brick(r1, r2, r3)
names(envBrick) <- c(0, 10, 30)

# Create test occurrences
set.seed(0)
longitude <- sample(extent(envBrick)[1]:extent(envBrick)[2],
                    size = 10, replace = FALSE)
set.seed(0)
latitude <- sample(extent(envBrick)[3]:extent(envBrick)[4],
                   size = 10, replace = FALSE)
set.seed(0)
depth <- sample(0:35, size = 10, replace = TRUE)
occurrences <- as.data.frame(cbind(longitude,latitude,depth))

# Test function
occSample3d <- xyzSample(occurrences, envBrick)

# How to use
occurrences$envtValue <- occSample3d

head(occurrences)


}
\keyword{dataPrep}
