% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{pointMap}
\alias{pointMap}
\title{Point mapping}
\usage{
pointMap(
  occs,
  spName,
  land = NA,
  ptCol = "#bd0026",
  landCol = "gray",
  waterCol = "steelblue",
  ...
)
}
\arguments{
\item{occs}{A \code{data.frame} with at least two columns
named "longitude" and "latitude" or that
can be coerced into this format.}

\item{spName}{A character string with the species
name to be used in the plot title.}

\item{land}{An optional coastline polygon shapefile
of type \code{sf} to provide geographic context for the
occurrence points.}

\item{ptCol}{Color for occurrence points on map}

\item{landCol}{Color for land on map}

\item{waterCol}{Color for water on map}

\item{...}{Additional optional arguments to pass to
\code{ggplot} initial plot object.}
}
\value{
A \code{data.frame} with two columns named "longitude"
and "latitude" or with names that were used when coercing
input data into this format.
}
\description{
A convenient wrapper around ggplot
to generate formatted occurrence point plots.
}
\details{
The meat of this function is a special-case wrapper
around \code{getDynamicAlphaHull} from the \code{rangeBuilder} package.
The function documented here is especially useful in cases where
one wants to automatically generate training regions that overlap
the international date line. Regions that exceed the line are cut
and pasted into the appropriate hemisphere instead of being
deleted.
}
\examples{
occs <- read.csv(system.file("extdata/Aphanopus_intermedius.csv",
                             package='voluModel'))
spName <- "Aphanopus intermedius"
pointMap(occs = occs, spName = spName, land = NA)

}
\seealso{
\code{\link[ggplot2:ggplot]{ggplot}}
}
\keyword{plotting}
